Remove nagios name to make the probe service agnostic
